[yoda] Fix config crash on warning
